+++ src/Components/main/ImageSlider.js
@@ -1,39 +1,39 @@
-// Import Swiper React components
-import { Swiper, SwiperSlide } from 'swiper/react';
+// // Import Swiper React components
+// import { Swiper, SwiperSlide } from 'swiper/react';
 
-// Import Swiper styles
-import 'swiper/css';
-import 'swiper/css/effect-coverflow';
-import 'swiper/css/pagination';
-import 'swiper/css/navigation';
+// // Import Swiper styles
+// import 'swiper/css';
+// import 'swiper/css/effect-coverflow';
+// import 'swiper/css/pagination';
+// import 'swiper/css/navigation';
 
-// import required modules
-import { EffectCoverflow, Pagination, Navigation } from 'swiper';
-export default function ImageSlider() {
-  return (
-    <>
-      <Swiper
-        effect={'coverflow'}
-        loop={true}
-        grabCursor={true}
-        centeredSlides={true}
-        slidesPerView={1}
-        slideToClickedSlide={true}
-        coverflowEffect={{
-          rotate: 0,
-          stretch: 100,
-          depth: 150,
-          modifier: 1.5,
-          slideShadows: false,
-          speed: 1000,
-        }}
-        pagination={{
-          clickable: true,
-        }}
-        modules={[EffectCoverflow, Pagination, Navigation]}
-        className="mySwiper"
-      >
-      </Swiper>
-    </>
-  );
-}
+// // import required modules
+// import { EffectCoverflow, Pagination, Navigation } from 'swiper';
+// export default function ImageSlider() {
+//   return (
+//     <>
+//       <Swiper
+//         effect={'coverflow'}
+//         loop={true}
+//         grabCursor={true}
+//         centeredSlides={true}
+//         slidesPerView={1}
+//         slideToClickedSlide={true}
+//         coverflowEffect={{
+//           rotate: 0,
+//           stretch: 100,
+//           depth: 150,
+//           modifier: 1.5,
+//           slideShadows: false,
+//           speed: 1000,
+//         }}
+//         pagination={{
+//           clickable: true,
+//         }}
+//         modules={[EffectCoverflow, Pagination, Navigation]}
+//         className="mySwiper"
+//       >
+//       </Swiper>
+//     </>
+//   );
+// }
