+++ src/redux/configstore.js
@@ -4,7 +4,7 @@ import Review from './Modules/PageModules/Review';
 import Detail from './Modules/PageModules/Detail';
 import Category from './Modules/PageModules/Category';
 
-const rootReducer = combineReducers({ Main, Review, Detail, Category });
+const rootReducer = combineReducers({ Main, Review, Detail });
 const store = configureStore({ reducer: rootReducer });
 
 export default store;
